version 3.06
What's new in release 3.06
Release date: 11 January 2009

  * Integrated 3270 (SYSG) console support (Roger Bowler, Jan Jaeger)
  * HMC DVD-RAM read/write support (Jan Jaeger)
  * 64-bit native version now supported on Mac OS X (Jay Maynard)
  * Ability to specify IFL, zIIP, and zAAP engine types (Roger Bowler, Jan
Jaeger, Ivan Warren)
  * Console-like message handling (David "Fish" Trout, Bernard van der Helm)
  * Tape automount CCW support (David "Fish" Trout)
  * CKD Locate Record Extended CCW (Greg Smith)
  * Support for FLEX-ES FakeTape tape images (David "Fish" Trout; FLEX-ES and
FakeTape are trademarks of Fundamental Software, Inc.)
  * More complete 3490 and 3590 tape support (David "Fish" Trout)
  * Solaris build support (Jeff Savit)
  * FreeBSD build support (Bjoern A. Zeeb)
  * Panel enhancements:

  * Display virtual storage in primary, secondary, and home space (Paul Leisy)
  * Display and modify PSW fields by panel command (Roger Bowler)
  * Modify control registers by panel command (Roger Bowler)
  * Specify IPL parameter by PARM operand (Ivan Warren)
  * New panel commands: automount, cmdtgt, ctc, herc, msghld, pscp, scp,
sfk (David "Fish" Trout, Bernard van der Helm)

+++ hdl.c
@@ -1,9 +1,12 @@
 /* HDL.C        (c) Copyright Jan Jaeger, 2003-2007                  */
 /*              Hercules Dynamic Loader                              */
 
-// $Id: hdl.c,v 1.53 2007/06/23 00:04:10 ivan Exp $
+// $Id: hdl.c,v 1.54 2008/11/23 23:29:44 rbowler Exp $
 //
 // $Log: hdl.c,v $
+// Revision 1.54  2008/11/23 23:29:44  rbowler
+// Fix win64 type conversion warnings in hdl.c
+//
 // Revision 1.53  2007/06/23 00:04:10  ivan
 // Update copyright notices to include current year (2007)
 //
@@ -143,7 +146,7 @@ static void * hdl_dlopen(char *filename, int flag _HDL_UNUSED)
 {
 char *fullname;
 void *ret;
-int fulllen = 0;
+size_t fulllen = 0;
 
     if(filename && *filename != '/' && *filename != '.')
     {
